pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sev0"
version = "0.1.0"
description = "AI-powered on-call triage agent"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "boto3>=1.35",
    "pyyaml>=6.0",
    "apscheduler>=3.10,<4.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "moto[cloudwatch,logs]>=5.0",
    "respx>=0.22",
]

[project.scripts]
sev0 = "sev0.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sev0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
